Convert environment to TypeScript
