+++ src/three-components/Model.ts
@@ -64,7 +64,7 @@ export default class Model extends Object3D {
     return !!this.modelContainer.children.length;
   }
 
-  applyEnvironmentMap(map: Texture) {
+  applyEnvironmentMap(map: Texture|null) {
     // Note that unlit models (using MeshBasicMaterial) should not apply
     // an environment map, even though `map` is the currently configured
     // environment map.
