+++ src/three-components/ModelScene.ts
@@ -66,11 +66,11 @@ export default class ModelScene extends Scene {
   private target = new Vector3();
   private canvas: HTMLCanvasElement;
   private renderer: Renderer;
-  private shadow: StaticShadow;
   private light: HemisphereLight;
   private shadowLight: DirectionalLight;
   private aspect: number;
 
+  public shadow: StaticShadow;
   public pivot: Object3D;
   public width: number;
   public height: number;
